Add files via upload
This project involves a few objects on the table. The user with the VR headset can reach these objects and interact with them as in a real world environment. Special objects called distance grabbers have been added. These objects can be grabbed from any place where the user is, regardless of the distance from the table.